requirements.txt
discord.py==2.5.2
google-genai==1.21.1
pillow==11.2.1